backend/requirement.txt
fastapi
uvicorn[standard]
torch
transformers
pydantic
requests
python-dotenv
gunicorn
